perf(subagents): reduce startup imports
