+++ .changeset/fast-subagents-entry.md
@@ -1,0 +1,5 @@
+---
+"@narumitw/pi-subagents": patch
+---
+
+Reduce idle startup work by separating Pi registration and read-only settings modules from backward-compatible utility exports and settings mutations.
