+++ packages/pi-subagents/src/settings.ts
@@ -15,32 +15,18 @@ import type {
 	SubagentTransportKind,
 } from "./agents/types.js";
 import { MAX_CONFIGURABLE_PARALLEL_TASKS } from "./limits.js";
-import {
-	type BlockingParallelLimitSettingsSnapshot,
-	buildBlockingParallelLimitSettingsSnapshot,
-	buildCompletionDeliverySettingsSnapshot,
-	buildConsultResourceSettingsSnapshot,
-	buildCwdPolicySettingsSnapshot,
-	buildDelegationWorkflowSettingsSnapshot,
-	buildStatefulLimitSettingsSnapshot,
-	buildStatefulTransportSettingsSnapshot,
-	buildSubagentSettingsSnapshot,
-	type CompletionDeliverySettingsSnapshot,
-	type ConsultResourceSettingsSnapshot,
-	type CwdPolicySettingsSnapshot,
-	type DelegationWorkflow,
-	type DelegationWorkflowSettingsSnapshot,
-	type InspectedSubagentSettingsDocument,
-	type StatefulLimitSettingsSnapshot,
-	type StatefulTransportSettingsSnapshot,
-	type SubagentSettingsSnapshot,
-} from "./settings/inspection.js";
+import type { DelegationWorkflow } from "./settings/inspection.js";
 import {
 	hasOwn,
 	isPlainObject,
 	isPositiveInteger,
 	normalizeSubagentSettings,
 } from "./settings/schema.js";
+import {
+	pathEntryExists,
+	resolveSubagentSettingsPaths,
+	SUBAGENT_SETTINGS_FILE,
+} from "./settings-reader.js";
 import {
 	isValidStatefulLimit,
 	resolveStatefulLimits,
@@ -56,26 +42,34 @@ export {
 	type ConsultResourceSettingsSnapshot,
 	type CwdPolicyFieldSnapshot,
 	type CwdPolicySettingsSnapshot,
+	consumeSubagentSettingsNotice,
 	DEFAULT_CONSULT_RESOURCE_POLICY,
 	DEFAULT_CONSULTATION_CWD_POLICY,
 	DEFAULT_DELEGATION_CWD_POLICY,
 	type DelegationWorkflow,
 	type DelegationWorkflowSettingsSnapshot,
+	hasOwn,
+	inspectBlockingParallelLimitSettings,
+	inspectCompletionDeliverySettings,
+	inspectConsultResourceSettings,
+	inspectCwdPolicySettings,
+	inspectDelegationWorkflowSettings,
+	inspectStatefulLimitSettings,
+	inspectStatefulTransportSettings,
+	inspectSubagentSettings,
+	normalizeAgentSettings,
+	normalizeSubagentSettings,
+	readSubagentSettings,
 	resolveBlockingMaxParallelTasks,
 	resolveDelegationWorkflow,
 	type StatefulLimitFieldSnapshot,
 	type StatefulLimitSettingsSnapshot,
 	type StatefulTransportSettingsSnapshot,
 	type SubagentSettingsSnapshot,
-} from "./settings/inspection.js";
-export {
-	hasOwn,
-	normalizeAgentSettings,
-	normalizeSubagentSettings,
-} from "./settings/schema.js";
+	subagentSettingsFilePath,
+} from "./settings-reader.js";
 
-const SETTINGS_FILE = "pi-subagents.json";
-const LEGACY_SETTINGS_FILE = "pi-subagents-config.json";
+const SETTINGS_FILE = SUBAGENT_SETTINGS_FILE;
 const require = createRequire(import.meta.url);
 
 const SETTINGS_LOCK_FS_ADAPTER = {
@@ -90,138 +84,9 @@ const SETTINGS_LOCK_FS_ADAPTER = {
 	utimes: fs.utimes,
 	utimesSync: fs.utimesSync,
 };
-let pendingSettingsNotice: string | undefined;
-
-function resolveSubagentSettingsPaths(): {
-	canonicalPath: string;
-	legacyPath: string;
-	activePath?: string;
-} {
-	const canonicalPath = path.join(getAgentDir(), SETTINGS_FILE);
-	const legacyPath = path.join(getAgentDir(), LEGACY_SETTINGS_FILE);
-	return {
-		canonicalPath,
-		legacyPath,
-		activePath: fs.existsSync(canonicalPath)
-			? canonicalPath
-			: fs.existsSync(legacyPath)
-				? legacyPath
-				: undefined,
-	};
-}
-
-export function readSubagentSettings(): SubagentSettings | undefined {
-	pendingSettingsNotice = undefined;
-	const { canonicalPath, legacyPath, activePath } = resolveSubagentSettingsPaths();
-	if (activePath === canonicalPath) {
-		const canonical = readSettingsFile(canonicalPath);
-		const notices: string[] = [];
-		if (!canonical) notices.push(`${SETTINGS_FILE} is invalid and was ignored.`);
-		if (fs.existsSync(legacyPath)) {
-			notices.push(`${LEGACY_SETTINGS_FILE} ignored because ${SETTINGS_FILE} takes precedence.`);
-		}
-		if (notices.length > 0) pendingSettingsNotice = notices.join("\n");
-		return canonical;
-	}
-	if (activePath === undefined) return undefined;
-	const legacy = readSettingsFile(legacyPath);
-	if (fs.existsSync(canonicalPath)) {
-		const canonical = readSettingsFile(canonicalPath);
-		pendingSettingsNotice = [
-			...(!canonical ? [`${SETTINGS_FILE} is invalid and was ignored.`] : []),
-			`${LEGACY_SETTINGS_FILE} ignored because ${SETTINGS_FILE} was created concurrently.`,
-		].join("\n");
-		return canonical;
-	}
-	if (!legacy) {
-		pendingSettingsNotice = `${LEGACY_SETTINGS_FILE} is invalid and was ignored.`;
-		return undefined;
-	}
-	pendingSettingsNotice = `Using legacy ${LEGACY_SETTINGS_FILE}; rename it to ${SETTINGS_FILE}. Future saves write ${SETTINGS_FILE} without modifying the legacy file.`;
-	return legacy;
-}
-
-export function consumeSubagentSettingsNotice() {
-	const notice = pendingSettingsNotice;
-	pendingSettingsNotice = undefined;
-	return notice;
-}
 
 export function saveSubagentConfig(settings: SubagentSettings): void {
 	writeSettingsObject(settings);
-}
-
-export function subagentSettingsFilePath(): string {
-	return path.join(getAgentDir(), SETTINGS_FILE);
-}
-
-function inspectSubagentSettingsDocument(): InspectedSubagentSettingsDocument {
-	const { canonicalPath, activePath } = resolveSubagentSettingsPaths();
-	if (activePath === undefined) return { path: canonicalPath };
-	const inspected = inspectSubagentSettingsPath(activePath);
-	return activePath !== canonicalPath && fs.existsSync(canonicalPath)
-		? inspectSubagentSettingsPath(canonicalPath)
-		: inspected;
-}
-
-function inspectSubagentSettingsPath(configPath: string): InspectedSubagentSettingsDocument {
-	const fileName = path.basename(configPath);
-	let contents: string;
-	try {
-		contents = fs.readFileSync(configPath, "utf8");
-	} catch (error) {
-		const code = (error as NodeJS.ErrnoException).code;
-		return {
-			path: configPath,
-			error: `${fileName} could not be read${code ? ` (${safeErrorCode(code)})` : ""}`,
-		};
-	}
-	let raw: unknown;
-	try {
-		raw = JSON.parse(contents);
-	} catch {
-		return { path: configPath, error: `${fileName} contains malformed JSON` };
-	}
-	const settings = normalizeSubagentSettings(raw);
-	if (!isPlainObject(raw) || !settings) {
-		return { path: configPath, error: `${fileName} is not a valid settings object` };
-	}
-	return { path: configPath, raw, settings };
-}
-
-export function inspectSubagentSettings(): SubagentSettingsSnapshot {
-	return buildSubagentSettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectConsultResourceSettings(): ConsultResourceSettingsSnapshot {
-	return buildConsultResourceSettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectCwdPolicySettings(): CwdPolicySettingsSnapshot {
-	return buildCwdPolicySettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectDelegationWorkflowSettings(): DelegationWorkflowSettingsSnapshot {
-	return buildDelegationWorkflowSettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectCompletionDeliverySettings(): CompletionDeliverySettingsSnapshot {
-	return buildCompletionDeliverySettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectStatefulTransportSettings(): StatefulTransportSettingsSnapshot {
-	return buildStatefulTransportSettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectBlockingParallelLimitSettings(): BlockingParallelLimitSettingsSnapshot {
-	return buildBlockingParallelLimitSettingsSnapshot(inspectSubagentSettingsDocument());
-}
-
-export function inspectStatefulLimitSettings(): StatefulLimitSettingsSnapshot {
-	return buildStatefulLimitSettingsSnapshot(
-		inspectSubagentSettingsDocument(),
-		subagentSettingsFilePath(),
-	);
 }
 
 export function updateDelegationWorkflowSetting(
@@ -531,36 +396,6 @@ function withSettingsMutationLock<T>(mutate: () => T): T {
 
 function sameStatefulLimits(left: StatefulLimits, right: StatefulLimits): boolean {
 	return STATEFUL_LIMIT_FIELDS.every((field) => left[field] === right[field]);
-}
-
-function pathEntryExists(filePath: string): boolean {
-	try {
-		fs.lstatSync(filePath);
-		return true;
-	} catch (error) {
-		if ((error as NodeJS.ErrnoException).code === "ENOENT") return false;
-		throw error;
-	}
-}
-
-function readSettingsFile(configPath: string): SubagentSettings | undefined {
-	return readSettingsSnapshot(configPath).settings;
-}
-
-function readSettingsSnapshot(configPath: string): {
-	settings?: SubagentSettings;
-	contents?: string;
-} {
-	try {
-		const contents = fs.readFileSync(configPath, "utf8");
-		return { settings: normalizeSubagentSettings(JSON.parse(contents)), contents };
-	} catch {
-		return {};
-	}
-}
-
-function safeErrorCode(value: string): string {
-	return value.replace(/[^A-Z0-9_-]/giu, "?").slice(0, 64);
 }
 
 export function uniqueToolNames(tools: string[]): string[] {
